requirements.txt
aiofiles==0.4.0
aiohttp==3.6.2
aioredis==1.3.1
alembic==1.3.1
astroid==2.3.3
async-timeout==3.0.1
asyncpg==0.20.0
attrs==19.3.0
autopep8==1.4.4
beautifulsoup4==4.8.2
blis==0.4.1
bs4==0.0.1
cachetools==3.1.1
catalogue==0.2.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cymem==2.0.3
databases==0.2.6
google-api-python-client==1.7.11
google-auth==1.7.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
gunicorn==20.0.4
h11==0.8.1
h2==3.1.1
hiredis==1.0.1
hpack==3.0.0
httpcore==0.3.0
httplib2==0.14.0
httptools==0.0.13
hyperframe==5.2.0
idna==2.8
importlib-metadata==1.3.0
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.4.2
Mako==1.1.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
multidict==4.6.1
murmurhash==1.0.2
numpy==1.18.0
oauthlib==3.1.0
plac==1.1.3
preshed==3.0.2
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
python-dateutil==2.8.1
python-editor==1.0.4
requests==2.22.0
requests-async==0.5.0
requests-oauthlib==1.3.0
rfc3986==1.3.2
rsa==4.0
sanic==19.9.0
Sanic-Cors==0.9.9.post3
Sanic-Plugins-Framework==0.8.2
sanic-session==0.5.6
six==1.13.0
soupsieve==1.9.5
spacy==2.2.3
SQLAlchemy==1.3.11
srsly==0.2.0
thinc==7.3.1
tqdm==4.41.1
typed-ast==1.4.0
ujson==1.35
uritemplate==3.0.0
urllib3==1.25.7
uvicorn==0.11.2
uvloop==0.14.0
wasabi==0.5.0
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
zipp==0.6.0
